+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "nf start",
     "start:dev": "DEBUG=floworky:* nf start --procfile Procfile.dev",
-    "start:dev:front": "node ./bin/client.js",
+    "start:dev:front": "node ./frontend/client.js",
     "webpack": "./node_modules/webpack-dev-server/bin/webpack-dev-server.js webpack-dev-server --progress --colors",
     "babel-node": "babel-node --presets=es2015",
     "test": "babel-node spec/run.js",
@@ -25,9 +25,10 @@
     "debug": "~2.2.0",
     "dotenv": "^2.0.0",
     "express": "~4.13.4",
+    "foreman": "^2.0.0",
     "fs": "^0.0.1-security",
     "jquery": "^3.1.1",
-    "foreman": "^2.0.0",
+    "knex": "^0.12.6",
     "morgan": "~1.7.0",
     "nodemon": "^1.11.0",
     "pg-promise": "^5.4.3",
